Added ESLint and Prettier
Created an .eslintrc file in the root (server) with all the global settings
and then added it to the extends inside client/package.json along with prettier/react.
This allowed vscode to format all the code on save using the same custom settings.

Ran ESLint and corrected all the errors